Add UI tests via PageObject pattern and environment for its running
Added 3 UI test via PageObject according to specification. Test are marked as UI to have an opportunity to run it with mark '-m UI'. Also, tests could be run in headless mode.
Tests could be run in parallel mode, on remote browser with Selenoid. Also, tests have UI report in Allure and they are covered with logger